+++ core/include/network/packet/PacketMoveEntity.hh
@@ -16,8 +16,8 @@ namespace network {
             PacketMoveEntity(uint32_t tick = 0,uint32_t eventId = 0, uint16_t entityId = 0, uint16_t vecx = 0, uint16_t vecy = 0);
             void setTick(uint32_t tick);
             uint32_t getTick() const;
-            void setEventId(uint16_t id);
-            uint16_t getEventId() const;
+            void setEventId(uint32_t id);
+            uint32_t getEventId() const;
             void setEntityId(uint16_t entityId);
             uint16_t getEntityId() const;
             void setVecX(uint16_t vecX);
